reactDRFGithub/requirements.txt
Django==1.11
djangorestframework
django-webpack-loader
